// src/assets/data/value.interface.ts
export interface Value {
    id?: string;
    name?: string;
    value?: number;
    product?: string;
    type?: string;  
    institution?: string;
    invDate?: string;
    initialValue?: number;
    currentValue?: number;
    description?: string;
    buyDate?: string;
    provider?: string;
    buyValue?: number;
    creditor?: string;
    debtDate?: number;
    interest?: number;
    mainValue?: number;
    remainingValue?: number;
}